+++ modules/Shmuelie.Git/Shmuelie.Git.psm1
@@ -70,6 +70,7 @@ Export-ModuleMember -Function @(
     'Get-GitStatusSummary',
     'Get-GitTag',
     'Remove-Branch',
+    'Set-Branch',
     'Format-GitStatusSegment',
     'Get-Worktrees',
     'Get-Branch',